pyproject.toml
[project]
name = "2025-09-22-gemini-code-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "httpx>=0.28.1",
    "openai>=1.108.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.13.2",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C", "N", "UP", "S", "B", "A", "C4", "ICN", "PIE", "T20", "SIM", "RUF"]
ignore = ["S101", "S603", "S607"]  # Allow assert statements and subprocess calls

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
